backend/requirements.txt
fastapi
uvicorn
python-multipart
faster-whisper
openai
supabase
python-dotenv
pydantic
pydantic-settings
httpx
gTTS
